requirements.txt
fastapi
uvicorn
sqlalchemy
pytz
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
fastapi-pagination